requirements.txt
click
pyperclip
pyaml
